Select category

+++ app/src/main/java/ru/antonc/budget/data/entities/Category.kt
@@ -5,7 +5,9 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = Category.TABLE_NAME)
 data class Category(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+
+    @PrimaryKey(autoGenerate = true)
+    val id: Long = 0,
 
     var name: String = ""
 ) {
